# perf: adds `Vector::InnerProdByElement` (E4 koalabear) in avx512