--- a/field/koalabear/extensions/e4_purego.go
+++ b/field/koalabear/extensions/e4_purego.go
@@ -52,3 +52,7 @@ func vectorButterfly_avx512(a, b *E4, N uint64) {
 func vectorButterflyPair_avx512(a *E4, N uint64) {
 	panic("vectorButterflyPair_avx512 is not implemented")
 }
+
+func vectorInnerProductByElement_avx512(res, a *E4, b *fr.Element, N uint64) {
+	panic("vectorInnerProductByElement_avx512 is not implemented")
+}