… field/babybear/poseidon2/poseidon2_amd64.s
… field/koalabear/extensions/e4_amd64.s
… field/koalabear/poseidon2/poseidon2_amd64.s