# Description

```
benchmark                                              old ns/op     new ns/op     delta
BenchmarkVectorOps/innerProductByElement_524288-96     4330025       342249        -92.10%
```

Also fixes an issue in avx512 code generation where `q` was passed incorrectly.